feat: atfx submodule
