+++ src/asamatfx/atfx/_binary_reader.py
