+++ tests/test_ec_resolution.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import odsbox.proto.ods_pb2 as ods
 import pytest
 
-from asamatfx import AtfxStore
+from asamatfx.atfx import AtfxStore
 
 DATA_DIR = Path(__file__).resolve().parent / "data" / "openatfx"
 ATFX_FILE = DATA_DIR / "example_toleratedIncorrect.atfx"
