+++ tests/test_model_builder.py
@@ -4,9 +4,9 @@ import xml.etree.ElementTree as ET
 
 import odsbox.proto.ods_pb2 as ods
 
-from asamatfx import AtfxStore
-from asamatfx._model_builder import detect_ods_version
-from asamatfx._xml_utils import _extract_ns
+from asamatfx.atfx import AtfxStore
+from asamatfx.atfx._model_builder import detect_ods_version
+from asamatfx.atfx._xml_utils import _extract_ns
 
 
 def test_simple_model_entity_count(simple_store):
